requirements-render.txt
# Core AI
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# API (needed for imports)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Database
sqlalchemy>=2.0.0
alembic>=1.14.0

# Telegram
python-telegram-bot>=21.0

# Search
tavily-python>=0.5.0

# Market Data
yfinance>=0.2.40

# Voice (for transcription)
openai>=1.50.0

# Utilities
httpx>=0.27.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0
pytz>=2024.2
numpy>=1.26.0
pandas>=2.2.0
jinja2>=3.1.0
gunicorn>=22.0.0
